Use docker-compose to orchestrate the services
Split the project out into 2 binaries.

See https://github.com/thockin/go-build-template/ for the new build
approach.

The first binary represents the frontend service which returns HTML to
the client. This service calls the backend upstream services to get
data, and aggregates that into the UI layer.

The second binary represents all of the upstream backend services. We
simulate many services by having multiple instances of this binary
running.

docker-compose is used to arrange these services together.

Tested on Mac.
